refactor(Domain): Bean Validation 삭제, 생성자 접근 제어자 protected로 변경, 전체 도메인 필드 Column annotation 추가
