+++ src/main/java/com/prgrms/mukvengers/domain/crewmember/model/CrewMember.java
@@ -14,17 +14,17 @@ import javax.persistence.ManyToOne;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
 
-import com.prgrms.mukvengers.domain.Crew;
-import com.prgrms.mukvengers.domain.User;
+import com.prgrms.mukvengers.domain.crew.model.Crew;
+import com.prgrms.mukvengers.domain.user.model.User;
 import com.prgrms.mukvengers.global.common.domain.BaseEntity;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@NoArgsConstructor(access = PROTECTED)
 @Entity
 @Getter
+@NoArgsConstructor(access = PROTECTED)
 @Where(clause = "deleted = false")
 @SQLDelete(sql = "UPDATE crew_member set deleted = true where id=?")
 public class CrewMember extends BaseEntity {
@@ -42,12 +42,14 @@ public class CrewMember extends BaseEntity {
 	@JoinColumn(name = "crew_id")
 	private Crew crew;
 
+	@Column(nullable = false)
 	private boolean blocked;
 
+	@Column(nullable = false)
 	private boolean ready;
 
 	@Builder
-	private CrewMember(User user, Crew crew, boolean blocked, boolean ready) {
+	protected CrewMember(User user, Crew crew, boolean blocked, boolean ready) {
 		this.user = user;
 		this.crew = crew;
 		this.blocked = blocked;
